Move the import and update tests in the correct package
